Implement minimal type parsing for let expressions
A problem with how rune parsed patterns was that object/struct
destructure patterns used a recursive call to parse the fields patterns
inside the object/struct. This means that `let x: type` was getting
yoinked as a field pattern even though it was not wrapped in braces.
This commit refactors the recursive call without making any changes to
the produced AST as it may conflict with work being done to implement
proc-macro support for rune.